+++ PropGrid/src/PropGrid.h
@@ -106,7 +106,7 @@
 	extern const struct TRationalNumber:public div_t{
 		TRationalNumber();
 
-		int operator*(short i) const;
+		int operator*(int i) const;
 		bool operator!=(int i) const;
 	} LogicalUnitScaleFactor;
